Se agrega session a los controllers.

+++ src/main/java/mx/edu/uttab/spring/controller/CalendarioCapacitacionController.java
@@ -1,5 +1,8 @@
 package mx.edu.uttab.spring.controller;
 
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpSession;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Controller;
@@ -31,45 +34,76 @@ public class CalendarioCapacitacionController {
 	}
 
 	@RequestMapping(value = "/calendarioscapacitaciones", method = RequestMethod.GET)
-	public String index(Model model) {
-		model.addAttribute("listCalendarioCapacitacion", this.calendarioCapacitacionService.listCalendarioCapacitacion());
-		return "calendarioCapacitacion/index";
+	public String index(Model model, HttpServletRequest request) {
+		HttpSession session = request.getSession();
+		if (session.getAttribute("nombre") != null && session.getAttribute("cve_usuario") != null) {
+			model.addAttribute("listCalendarioCapacitacion",
+					this.calendarioCapacitacionService.listCalendarioCapacitacion());
+			return "calendarioCapacitacion/index";
+		} else {
+			return "redirect:/";
+		}
+
 	}
 
 	@RequestMapping(value = "/calendarioscapacitaciones/new", method = RequestMethod.GET)
-	public String create(Model model) {
-		model.addAttribute("calendarioCapacitacion", new CalendarioCapacitacion());
-		model.addAttribute("listCapacitacion", this.capacitacionService.listCapacitacion());
-		return "calendarioCapacitacion/create";
+	public String create(Model model, HttpServletRequest request) {
+		HttpSession session = request.getSession();
+		if (session.getAttribute("nombre") != null && session.getAttribute("cve_usuario") != null) {
+			model.addAttribute("calendarioCapacitacion", new CalendarioCapacitacion());
+			model.addAttribute("listCapacitacion", this.capacitacionService.listCapacitacion());
+			return "calendarioCapacitacion/create";
+		} else {
+			return "redirect:/";
+		}
+
 	}
 
 	@RequestMapping(value = "/calendarioscapacitaciones/create", method = RequestMethod.POST)
-	public String store(@ModelAttribute("calendarioCapacitacion") CalendarioCapacitacion cc) {
+	public String store(@ModelAttribute("calendarioCapacitacion") CalendarioCapacitacion cc,
+			HttpServletRequest request) {
+		HttpSession session = request.getSession();
+		if (session.getAttribute("nombre") != null && session.getAttribute("cve_usuario") != null) {
+			if (cc.getId() == 0) {
+				// new person, add it
+				this.calendarioCapacitacionService.addCalendarioCapacitacion(cc);
+				;
+			} else {
+				// existing person, call update
+				this.calendarioCapacitacionService.updateCalendarioCapacitacion(cc);
+			}
 
-		if (cc.getId() == 0) {
-			// new person, add it
-			this.calendarioCapacitacionService.addCalendarioCapacitacion(cc);
-			;
+			return "redirect:/calendarioscapacitaciones";
 		} else {
-			// existing person, call update
-			this.calendarioCapacitacionService.updateCalendarioCapacitacion(cc);
+			return "redirect:/";
 		}
-
-		return "redirect:/calendarioscapacitaciones";
 
 	}
 
 	@RequestMapping("/calendarioscapacitaciones/{id}/edit")
-	public String edit(@PathVariable("id") int id, Model model) {
-		model.addAttribute("calendarioCapacitacion", this.calendarioCapacitacionService.getCalendarioCapacitacionById(id));
-		model.addAttribute("listCapacitacion", this.capacitacionService.listCapacitacion());
-		return "calendarioCapacitacion/edit";
+	public String edit(@PathVariable("id") int id, Model model, HttpServletRequest request) {
+		HttpSession session = request.getSession();
+		if (session.getAttribute("nombre") != null && session.getAttribute("cve_usuario") != null) {
+			model.addAttribute("calendarioCapacitacion",
+					this.calendarioCapacitacionService.getCalendarioCapacitacionById(id));
+			model.addAttribute("listCapacitacion", this.capacitacionService.listCapacitacion());
+			return "calendarioCapacitacion/edit";
+		} else {
+			return "redirect:/";
+		}
+
 	}
 
 	@RequestMapping("/calendarioscapacitaciones/{id}/destroy")
-	public String destroy(@PathVariable("id") int id) {
-		this.calendarioCapacitacionService.removeCalendarioCapacitacion(id);
-		return "redirect:/calendarioscapacitaciones";
+	public String destroy(@PathVariable("id") int id, HttpServletRequest request) {
+		HttpSession session = request.getSession();
+		if (session.getAttribute("nombre") != null && session.getAttribute("cve_usuario") != null) {
+			this.calendarioCapacitacionService.removeCalendarioCapacitacion(id);
+			return "redirect:/calendarioscapacitaciones";
+		} else {
+			return "redirect:/";
+		}
+
 	}
 
 }
