+++ src/main/java/mx/edu/uttab/spring/controller/EmpresaController.java
@@ -5,6 +5,9 @@
  */
 package mx.edu.uttab.spring.controller;
 
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpSession;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Controller;
@@ -36,45 +39,73 @@ public class EmpresaController {
 	}
 
 	@RequestMapping(value = "/empresas", method = RequestMethod.GET)
-	public String index(Model model) {
-		model.addAttribute("listEmpresa", this.empresaService.listEmpresa());
-		return "empresas/index";
+	public String index(Model model, HttpServletRequest request) {
+		HttpSession session = request.getSession();
+		if (session.getAttribute("nombre") != null && session.getAttribute("cve_usuario") != null) {
+			model.addAttribute("listEmpresa", this.empresaService.listEmpresa());
+			return "empresas/index";
+		} else {
+			return "redirect:/";
+		}
+
 	}
 
 	@RequestMapping(value = "/empresas/new", method = RequestMethod.GET)
-	public String create(Model model) {
-		model.addAttribute("empresa", new Empresa());
-		model.addAttribute("listSectorProductivo", this.sectorProductivoService.listSectorProductivos());
-		return "empresas/create";
+	public String create(Model model, HttpServletRequest request) {
+		HttpSession session = request.getSession();
+		if (session.getAttribute("nombre") != null && session.getAttribute("cve_usuario") != null) {
+			model.addAttribute("empresa", new Empresa());
+			model.addAttribute("listSectorProductivo", this.sectorProductivoService.listSectorProductivos());
+			return "empresas/create";
+		} else {
+			return "redirect:/";
+		}
+
 	}
 
 	@RequestMapping(value = "/empresas/create", method = RequestMethod.POST)
-	public String store(@ModelAttribute("empresa") Empresa e) {
+	public String store(@ModelAttribute("empresa") Empresa e, HttpServletRequest request) {
+		HttpSession session = request.getSession();
+		if (session.getAttribute("nombre") != null && session.getAttribute("cve_usuario") != null) {
+			if (e.getId() == 0) {
+				// new person, add it
+				this.empresaService.addEmpresa(e);
+				;
+			} else {
+				// existing person, call update
+				this.empresaService.updateEmpresa(e);
+			}
 
-		if (e.getId() == 0) {
-			// new person, add it
-			this.empresaService.addEmpresa(e);
-			;
+			return "redirect:/empresas";
 		} else {
-			// existing person, call update
-			this.empresaService.updateEmpresa(e);
+			return "redirect:/";
 		}
-
-		return "redirect:/empresas";
 
 	}
 
 	@RequestMapping("/empresas/{id}/edit")
-	public String edit(@PathVariable("id") int id, Model model) {
-		model.addAttribute("empresa", this.empresaService.getEmpresaById(id));
-		model.addAttribute("listSectorProductivo", this.sectorProductivoService.listSectorProductivos());
-		return "empresas/edit";
+	public String edit(@PathVariable("id") int id, Model model, HttpServletRequest request) {
+		HttpSession session = request.getSession();
+		if (session.getAttribute("nombre") != null && session.getAttribute("cve_usuario") != null) {
+			model.addAttribute("empresa", this.empresaService.getEmpresaById(id));
+			model.addAttribute("listSectorProductivo", this.sectorProductivoService.listSectorProductivos());
+			return "empresas/edit";
+		} else {
+			return "redirect:/";
+		}
+
 	}
 
 	@RequestMapping("/empresas/{id}/destroy")
-	public String destroy(@PathVariable("id") int id) {
-		this.empresaService.removeEmpresa(id);
-		return "redirect:/empresas";
+	public String destroy(@PathVariable("id") int id, HttpServletRequest request) {
+		HttpSession session = request.getSession();
+		if (session.getAttribute("nombre") != null && session.getAttribute("cve_usuario") != null) {
+			this.empresaService.removeEmpresa(id);
+			return "redirect:/empresas";
+		} else {
+			return "redirect:/";
+		}
+
 	}
 
 }
